# frontend/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "okta": []
    }
  ],
  "paths": {
    "/me": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Get current authenticated user",
        "description": "Returns the currently authenticated user's details. If the user doesn't exist in the system, it tries to sync from Auth0.",
        "operationId": "getUser",
        "responses": {
          "200": {
            "description": "Authenticated user info",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "User not authenticated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "sub": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "id",
          "sub"
        ]
      },
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "okta": {
        "type": "oauth2",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://dev-l5aemj1026u4dqia.us.auth0.com/authorize",
            "tokenUrl": "https://dev-l5aemj1026u4dqia.us.auth0.com/oauth/token",
            "scopes": {
              "openid": "openid",
              "email": "email",
              "profile": "profile"
            }
          }
        }
      }
    }
  }
}